+++ tests/dummy/app/items/item/edit/controller.js
@@ -28,7 +28,24 @@ export default Ember.Controller.extend({
     cancel: function () {
       // return to the item list
       this.transitionToRoute('items.index');
+    },
+    removeResource (resource) {
+      const item = this.get('model.item');
+      return this.get('itemsService').removeResource(item.id, item.owner, resource.resource);
+    },
+    uploadFile (file) {
+      const item = this.get('model.item');
+      return this.get('itemsService').uploadResource(item.id, item.owner, file);
+    },
+    sendJson (jsonObj) {
+      const item = this.get('model.item');
+      return this.get('itemsService').addResource(item.id, item.owner, 'theme.json', JSON.stringify(jsonObj));
+    },
+    fetchResources(){
+      const item = this.get('model.item');
+      return this.get('itemsService').getResources(item.id);
     }
+
   }
 
 });